Refactor main file, add support for postgres certificate authentication
