+++ backend/internal/sys/config/conf_database.go
@@ -20,4 +20,7 @@ type Database struct {
 	SslMode          string `yaml:"ssl_mode"`
 	SqlitePath       string `yaml:"sqlite_path"        conf:"default:./.data/homebox.db?_pragma=busy_timeout=999&_pragma=journal_mode=WAL&_fk=1&_time_format=sqlite"`
 	PubSubConnString string `yaml:"pubsub_conn_string" conf:"default:mem://{{ .Topic }}"`
+	SslRootCert      string `yaml:"sslrootcert"`
+	SslCert          string `yaml:"sslcert"`
+	SslKey           string `yaml:"sslkey"`
 }
